quandoo_challenge/quandoo_challenge/scraper_app/requirements.txt
pandas==1.3.3
sqlalchemy==1.4.22
psycopg2-binary
requests==2.31.0
beautifulsoup4==4.12.2